engine/pyproject.toml
[project]
name = "memra"
version = "1.0.0"
description = "Memra 1.0 — context engine with persistent memory for AI agents"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "sse-starlette>=2.0.0",
    "anthropic>=0.49.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
